*** empty log message ***

+++ lib/data.h
@@ -8,7 +8,7 @@
  * See the file "COPYING" for information on usage and redistribution
  * of this file, and for a DISCLAIMER OF ALL WARRANTIES.
  *
- * @(#) $Id: data.h,v 1.65 2000/07/06 14:40:42 strauss Exp $
+ * @(#) $Id: data.h,v 1.66 2000/10/19 16:14:40 strauss Exp $
  */
 
 #ifndef _DATA_H
@@ -22,7 +22,7 @@
 
 
 typedef struct List {
-    SmiElement      export; /*  */
+    SmiElement      export;
     void	    *ptr;
     struct List	    *nextPtr;
 } List;
@@ -41,7 +41,7 @@ typedef enum Kind {
 
 
 
-typedef unsigned int ParserFlags;
+typedef unsigned short ParserFlags;
 typedef unsigned short ModuleFlags;
 typedef unsigned short ObjectFlags;
 typedef unsigned short NodeFlags;
